✨ split contact form and social links into two components
